backend/requirements.txt
django
djangorestframework
django-cors-headers
numpy
scikit-learn
tensorflow
twilio